Sources

Source

Chatter.Source — Type.

A source of communication.

Note that Source is abstract, and that its subtypes have the responsibility of implementing listen and deliver such that their behaviour is as described below.

source

Chatter.listen — Function.

listen(source::Source, channel::Channel{Message}) -> Task

Listen for messages and put! them into channel.

Arguments

source::Source: Source of communication.

channel::Channel{Message}: Channel for passing messages.

Returns the listening Task.

source

Chatter.deliver — Function.

deliver(source::Source, msg::AbstractString) -> Int

Deliver a message via source.

Arguments

source::Source: Source of communication.

msg::AbstractString: Message to be delivered.

Returns the number of bytes sent.

source

IRC

Chatter.IRC — Type.

IRC(;
    server::AbstractString=error("No server given"),
    channel::AbstractString=error("No channel given"),
    port::Int=6667,
    nick::AbstractString=randstring(16),
)

Initialize a new IRC channel source.

Arguments

server::AbstractString=error("No server given"): Server to connect to.

channel::AbstractString=error("No channel given"): Channel to join.

port::Int=6667: Port to connect via.

nick::AbstractString=randstring(16): Nickname to assume.

source

Chatter.IRCMessage — Type.

IRCMessage(body::AbstractString, author::AbstractString) -> IRCMessage

Create a new IRC message.

Arguments

body::AbstractString: Message contents.

author::AbstractString: Author of the message.

source

Subreddit

Chatter.Subreddit — Type.

Subreddit(sub::AbstractString; feed::AbstractString="new")

Initialize a new subreddit source.

Arguments

sub::AbstractString: Subreddit to monitor.

Keyword Arguments

feed::AbstractString="new": Retrieve new posts or hot posts.

source

Chatter.SubredditMessage — Type.

SubredditMessage(title::AbstractString, link::AbstractString) -> SubredditMessage

Create a new subreddit message.

Arguments

title::AbstractString: Title of the post.

link::AbstractString: Shortened URL.

source